+++ pynamodb/attributes.py
@@ -438,39 +438,6 @@ class JSONAttribute(Attribute):
         return json.loads(value, strict=False)
 
 
-class LegacyBooleanAttribute(Attribute):
-    """
-    A class for legacy boolean attributes
-
-    Previous versions of this library serialized bools as numbers.
-    This class allows you to continue to use that functionality.
-    """
-
-    attr_type = NUMBER
-
-    def serialize(self, value):
-        if value is None:
-            return None
-        elif value:
-            return json.dumps(1)
-        else:
-            return json.dumps(0)
-
-    def deserialize(self, value):
-        return bool(json.loads(value))
-
-    def get_value(self, value):
-        # we need this for the period in which you are upgrading
-        # you can switch all BooleanAttributes to LegacyBooleanAttributes
-        # this can read both but serializes as Numbers
-        # once you've transitioned, you can then switch back to
-        # BooleanAttribute and it will serialize the new fancy way
-        value_to_deserialize = super(LegacyBooleanAttribute, self).get_value(value)
-        if value_to_deserialize is None:
-            value_to_deserialize = json.dumps(value.get(BOOLEAN, 0))
-        return value_to_deserialize
-
-
 class BooleanAttribute(Attribute):
     """
     A class for boolean attributes
@@ -487,14 +454,6 @@ class BooleanAttribute(Attribute):
 
     def deserialize(self, value):
         return bool(value)
-
-    def get_value(self, value):
-        # we need this for legacy compatibility.
-        # previously, BOOL was serialized as N
-        value_to_deserialize = super(BooleanAttribute, self).get_value(value)
-        if value_to_deserialize is None:
-            value_to_deserialize = json.loads(value.get(NUMBER_SHORT, '0'))
-        return value_to_deserialize
 
 
 class NumberSetAttribute(SetMixin, Attribute):