+++ pynamodb/tests/data.py
@@ -1099,38 +1099,9 @@ COMPLEX_MODEL_TABLE_DATA = {
      }
 }
 
-BOOLEAN_CONVERSION_MODEL_TABLE_DATA_OLD_STYLE = {
+BOOLEAN_MODEL_TABLE_DATA = {
     'Table': {
-        'ItemCount': 0, 'TableName': 'BooleanConversionTable',
-        'ProvisionedThroughput': {
-            'ReadCapacityUnits': 2,
-            'WriteCapacityUnits': 2,
-            'NumberOfDecreasesToday': 0
-        },
-        'CreationDateTime': 1391471876.86,
-        'TableStatus': 'ACTIVE',
-        'AttributeDefinitions': [
-            {
-                'AttributeName': 'user_name',
-                'AttributeType': 'S'
-            },
-            {
-                'AttributeName': 'is_human',
-                'AttributeType': 'N'
-            }
-        ],
-        'KeySchema': [
-            {
-                'AttributeName': 'user_name', 'KeyType': 'HASH'
-            }
-        ],
-        'TableSizeBytes': 0
-    }
-}
-
-BOOLEAN_CONVERSION_MODEL_TABLE_DATA = {
-    'Table': {
-        'ItemCount': 0, 'TableName': 'BooleanConversionTable',
+        'ItemCount': 0, 'TableName': 'BooleanTable',
         'ProvisionedThroughput': {
             'ReadCapacityUnits': 2,
             'WriteCapacityUnits': 2,
@@ -1157,29 +1128,7 @@ BOOLEAN_CONVERSION_MODEL_TABLE_DATA = {
     }
 }
 
-BOOLEAN_CONVERSION_MODEL_OLD_STYLE_TRUE_ITEM_DATA = {
-    'Item': {
-        'user_name': {
-            'S': 'justin'
-        },
-        'is_human': {
-            'N': '1'
-        }
-    }
-}
-
-BOOLEAN_CONVERSION_MODEL_OLD_STYLE_FALSE_ITEM_DATA = {
-    'Item': {
-        'user_name': {
-            'S': 'alf'
-        },
-        'is_human': {
-            'N': '0'
-        }
-    }
-}
-
-BOOLEAN_CONVERSION_MODEL_NEW_STYLE_TRUE_ITEM_DATA = {
+BOOLEAN_MODEL_TRUE_ITEM_DATA = {
     'Item': {
         'user_name': {
             'S': 'justin'
@@ -1190,7 +1139,7 @@ BOOLEAN_CONVERSION_MODEL_NEW_STYLE_TRUE_ITEM_DATA = {
     }
 }
 
-BOOLEAN_CONVERSION_MODEL_NEW_STYLE_FALSE_ITEM_DATA = {
+BOOLEAN_MODEL_FALSE_ITEM_DATA = {
     'Item': {
         'user_name': {
             'S': 'alf'